format cub3d.c - format view rendering functions
