+++ src/vector_funcs.c
@@ -6,7 +6,7 @@
 /*   By: leferrei <leferrei@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/30 17:06:09 by leferrei          #+#    #+#             */
-/*   Updated: 2023/02/07 18:03:31 by leferrei         ###   ########.fr       */
+/*   Updated: 2023/02/08 15:47:50 by leferrei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,37 +42,35 @@ int	check_collision(t_vec *pos, t_vec *mov_vec, t_vars **vars)
 			+ (int)new_pos[0]] == '1');
 }
 
+void	set_width_vects(t_vec *width_dir, t_vec *width_pos
+	, t_vars **vars, double angle)
+{
+	set_vect_to_vect(width_dir, &(*vars)->p_vec->p_dir);
+	scale_vect(width_dir, 0.166);
+	rotate_vec(width_dir, angle);
+	set_vect_to_vect(width_pos, &(*vars)->p_vec->p_pos);
+	add_vect(width_pos, *width_dir);
+}
+
 int	is_movement_coliding(t_vec *pos, t_vec *mov_vec, t_vars **vars)
 {
-	int		result;
 	t_vec	check_mov_vector;
 	t_vec	transf_vec;
 	t_vec	width_pos;
 	t_vec	width_dir;
 
 	set_vect_to_vect(&check_mov_vector, mov_vec);
-	result = check_collision(pos, &check_mov_vector, vars);
-	if (result)
+	if (check_collision(pos, &check_mov_vector, vars))
 		return (1);
 	set_vect_to_vect(&transf_vec, &check_mov_vector);
 	set_vect_to_vect(&check_mov_vector, mov_vec);
-	set_vect_to_vect(&width_dir, &(*vars)->p_vec->p_dir);
-	scale_vect(&width_dir, 0.166);
-	rotate_vec(&width_dir, 45);
-	set_vect_to_vect(&width_pos, &(*vars)->p_vec->p_pos);
-	add_vect(&width_pos, width_dir);
+	set_width_vects(&width_dir, &width_pos, vars, 45);
 	set_vect_to_vect(&check_mov_vector, mov_vec);
-	result = check_collision(&width_pos, &check_mov_vector, vars);
-	if (result)
+	if (check_collision(&width_pos, &check_mov_vector, vars))
 		return (1);
-	set_vect_to_vect(&width_dir, &(*vars)->p_vec->p_dir);
-	scale_vect(&width_dir, 0.166);
-	rotate_vec(&width_dir, -45);
-	set_vect_to_vect(&width_pos, &(*vars)->p_vec->p_pos);
-	add_vect(&width_pos, width_dir);
+	set_width_vects(&width_dir, &width_pos, vars, -45);
 	set_vect_to_vect(&check_mov_vector, mov_vec);
-	result = check_collision(&width_pos, &check_mov_vector, vars);
-	if (result)
+	if (check_collision(&width_pos, &check_mov_vector, vars))
 		return (1);
 	set_vect_to_vect(mov_vec, &transf_vec);
 	return (0);
