+++ src/vector_math.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   vector_math.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: drobert- <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: leferrei <leferrei@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/07 18:43:25 by drobert-          #+#    #+#             */
-/*   Updated: 2023/02/07 18:43:26 by drobert-         ###   ########.fr       */
+/*   Updated: 2023/02/08 15:36:01 by leferrei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
